[refactor] Move test context manager to pylint.testutils
